Slack bot server updates
